+++ PulseServer/src/org/java_websocket/drafts/Draft_10.java
@@ -113,7 +113,7 @@ public class Draft_10 extends Draft {
 		assert (payloadlengthbytes.length == sizebytes);
 
 		if (sizebytes == 1) {
-			buf.put((byte) ((byte) payloadlengthbytes[0] | (mask ? (byte) -128
+			buf.put((byte) (payloadlengthbytes[0] | (mask ? (byte) -128
 					: 0)));
 		} else if (sizebytes == 2) {
 			buf.put((byte) ((byte) 126 | (mask ? (byte) -128 : 0)));
@@ -408,7 +408,7 @@ public class Draft_10 extends Draft {
 			byte[] maskskey = new byte[4];
 			buffer.get(maskskey);
 			for (int i = 0; i < payloadlength; i++) {
-				payload.put((byte) ((byte) buffer.get( /* payloadstart + i */) ^ (byte) maskskey[i % 4]));
+				payload.put((byte) (buffer.get( /* payloadstart + i */) ^ maskskey[i % 4]));
 			}
 		} else {
 			payload.put(buffer.array(), buffer.position(), payload.limit());
